Initial commit
Layout for the project. The goal is to use a single set of tests (will
be contained in test/support/all_tests.rb) and run them against the
two possible implementations (lib/kwic/v1.rb and lib/kwic/v2.rb). This
sets that up and adds a few rake tasks to help with running the tests